+++ selfdrive/car/tesla/pedal_calibrator/info.sh
@@ -1,0 +1,8 @@
+#! /bin/sh
+echo "================================="
+echo "Welcome to the Pedal Calibrator"
+echo "================================="
+echo "This app will calibrate the pedal interceptor so you can use OpenPilot on preAP Tesla Model S."
+echo "Please press the brake pedal, put the car in neutral(N) then hit Calibrate to start the process."
+echo "NOTE: KEEP BRAKE PEDAL PRESSED UNTIL PATCH PROCESS IS COMPLETE"
+echo " "